assuage 15000 warnings in G++
and other things.

+++ src/Basescape/DismantleFacilityState.h
@@ -43,7 +43,7 @@ class DismantleFacilityState
 {
 
 private:
-	int _refund;
+	int64_t _refund;
 
 	Base* _base;
 	const BaseFacility* _fac;
